# Startup performance improvements (less reflection)