--- a/Documentation/release-notes/4302.md
+++ b/Documentation/release-notes/4302.md
@@ -1,0 +1,11 @@
+### App startup performance
+
+  * [GitHub PR 4302](https://github.com/xamarin/xamarin-android/pull/4302):
+    Avoid unneeded calls to `GetCustomAttribute()` during app startup for the
+    common case where an app has no types decorated with the
+    `[JniAddNativeMethodRegistration]` attribute.  Additionally, instead of
+    using a managed method to propagate uncaught exceptions from Java, use a
+    Java method that calls into the unmanaged Xamarin.Android runtime.  These
+    changes reduced the time to display the first screen of a small test
+    Xamarin.Forms app from about 730 milliseconds to about 700 milliseconds for
+    a Release configuration build on a Google Pixel 3 XL device.