--- a/src/Xamarin.Android.Build.Tasks/Utilities/ApplicationConfigTaskState.cs
+++ b/src/Xamarin.Android.Build.Tasks/Utilities/ApplicationConfigTaskState.cs
@@ -1,0 +1,9 @@
+namespace Xamarin.Android.Tasks
+{
+	class ApplicationConfigTaskState
+	{
+		public const string RegisterTaskObjectKey = "Xamarin.Android.Tasks.ApplicationConfigTaskState";
+
+		public bool JniAddNativeMethodRegistrationAttributePresent { get; set; } = false;
+	}
+}