--- a/src/java-runtime/java/mono/android/Runtime.java
+++ b/src/java-runtime/java/mono/android/Runtime.java
@@ -1,5 +1,8 @@
 package mono.android;
 
+import java.lang.Thread;
+import java.lang.Throwable;
+
 public class Runtime {
 	static java.lang.Class java_lang_Class = java.lang.Class.class;;
 	static java.lang.Class java_lang_System = java.lang.System.class;
@@ -7,8 +10,8 @@ public class Runtime {
 	static java.lang.Class mono_android_IGCUserPeer = mono.android.IGCUserPeer.class;
 	static java.lang.Class mono_android_GCUserPeer = mono.android.GCUserPeer.class;
 
-	private Runtime ()
-	{
+	static {
+		Thread.setDefaultUncaughtExceptionHandler (new XamarinUncaughtExceptionHandler (Thread.getDefaultUncaughtExceptionHandler ()));
 	}
 
 	public static native void init (String lang, String[] runtimeApks, String runtimeDataDir, String[] appDirs, ClassLoader loader, String[] externalStorageDirs, String[] assemblies, String packageName, int apiLevel, String[] environmentVariables);
@@ -20,4 +23,22 @@ public class Runtime {
 	public static native void switchToContext (int contextID);
 	public static native void destroyContexts (int[] contextIDs);
 	public static native void propagateUncaughtException (Thread javaThread, Throwable javaException);
+}
+
+final class XamarinUncaughtExceptionHandler implements Thread.UncaughtExceptionHandler {
+	Thread.UncaughtExceptionHandler defaultHandler;
+
+	public XamarinUncaughtExceptionHandler (Thread.UncaughtExceptionHandler previousHandler)
+	{
+		defaultHandler = previousHandler;
+	}
+
+	@Override
+	public final void uncaughtException (Thread t, Throwable e)
+	{
+		Runtime.propagateUncaughtException (t, e);
+
+		if (defaultHandler != null)
+			defaultHandler.uncaughtException (t, e);
+	}
 }