--- a/src/monodroid/jni/xamarin-app.hh
+++ b/src/monodroid/jni/xamarin-app.hh
@@ -65,6 +65,7 @@ struct ApplicationConfig
 	bool is_a_bundled_app;
 	bool broken_exception_transitions;
 	bool instant_run_enabled;
+	bool jni_add_native_method_registration_attribute_present;
 	uint8_t bound_exception_type;
 	uint32_t package_naming_policy;
 	uint32_t environment_variable_count;