@grendello 2020-02-24
@radekdoulik I think `Activity Displayed` is simply not a good way to measure our performance. This figure depends on a number of factors on the Android side - it's the time it takes Android to render the activity surface, make it visible etc etc. It can be affected by a multitude of factors outside of our control. I'm working on a more reliable "entire startup" performance measurement now.