@radekdoulik 2020-02-24
It was quite reliable for me so far. It would be interesting to compare the Activity Displayed times with video recording to see how much does it correlate to the wall clock.