@grendello 2020-02-24
it appears to be heavily influenced by what the device/emulator is doing at the time. From what I saw in a quick test the difference of time between the end of `OnResume` and `Activity Displayed` was around 200-250ms (over 10 runs)