@radekdoulik 2020-02-25
I was thinking about the times here and my guess is that because we are postponing the uncaught handler, we are not doing any jni type registration in runtime init anymore. Thus all the JITing of the registration code is delayed between runtime init and activity displayed? That would explain the large gain in runtime init and small difference around activity displayed.

This guess might be confirmed by measuring jit times if needed (https://github.com/xamarin/xamarin-android/blob/master/Documentation/guides/profiling.md#profiling-the-jit-compiler)