@grendello 2020-02-25
@radekdoulik yes, less JIT time plays a large role here, another is JNI invocations (much less impact than JIT, though). It is possible that some of the JIT load shifted to the "end" of startup, in this case to the rendering phase.