@brendanzagaeski 2020-02-29
*Draft release notes*

Here's a candidate release note for this change. Feel free to edit as desired and add it to this pull request. (Additional info about preparing draft release notes can now be found in the [release notes README](https://github.com/xamarin/xamarin-android/blob/master/Documentation/release-notes/README.md).)

```
### App startup performance

  * [GitHub PR 3729](https://github.com/xamarin/xamarin-android/pull/3729):
    Avoid unneeded calls to `GetCustomAttribute()` during app startup for the
    common case where an app has no types decorated with the
    `[JniAddNativeMethodRegistration]` attribute.  Additionally, instead of
    using a managed method to propagate uncaught exceptions from Java, use a
    Java method that calls into the unmanaged Xamarin.Android runtime.  These
    changes reduced the time to display the first screen of a small test
    Xamarin.Forms app from about 730 milliseconds to about 700 milliseconds for
    a Release configuration build on a Google Pixel 3 XL device.
```