@grendello 2020-03-06
Hmm, the tests failed on jenkins again, however locally they worked fine:

```
~$ adb logcat -c; adb shell am instrument -w Mono.Android_TestsMultiDex/xamarin.android.runtimetests.TestInstrumentation; adb logcat -d > /tmp/log.txt
INSTRUMENTATION_RESULT: failed=0
INSTRUMENTATION_RESULT: inconclusive=0
INSTRUMENTATION_RESULT: nunit2-results-path=/storage/emulated/0/Android/data/Mono.Android_TestsMultiDex/files/Documents/TestResults.xml
INSTRUMENTATION_RESULT: passed=769
INSTRUMENTATION_RESULT: run=771
INSTRUMENTATION_RESULT: skipped=2
INSTRUMENTATION_CODE: -1
```